auth_app/requirements.txt
flask==2.2.3
python-dotenv==1.0.0
sqlalchemy==2.0.9
gevent==22.10.2
flask_sqlalchemy==3.0.3
flasgger==0.9.5
pydantic==1.9.0
psycopg2-binary==2.9.4
alembic==1.10.3
flask_jwt_extended==4.4.4
redis==4.5.4
flask-swagger-ui==4.11.1
pytest==7.3.1
requests==2.29.0
user-agents==2.2.0
opentelemetry-api==1.10.0
opentelemetry-sdk==1.10.0
opentelemetry-instrumentation-flask==0.29b1
opentelemetry-exporter-jaeger==1.10.0
